[mlir][sparse] Change DimLevelType Enum
Update DimLevelType to use lower 8 bits for storage formats and the higher 4 bits to store level properties. Treat CompressedWithHi and TwoOutOfFour as properties instead of formats.

Example: Compressed is 0b00000000_00000010 and CompressedWithHi is 0b00000100_00000010. It indicates that CompressedWithHi's format is Compressed and its property is High.

+++ mlir/include/mlir/ExecutionEngine/SparseTensor/Storage.h
@@ -683,7 +683,7 @@ private:
     if (isDenseDLT(dlt))
       return parentSz * getLvlSizes()[l];
     MLIR_SPARSETENSOR_FATAL("unsupported level type: %d\n",
-                            static_cast<uint8_t>(dlt));
+                            static_cast<uint16_t>(dlt));
   }
 
   /// Initializes sparse tensor storage scheme from a memory-resident sparse
